Create the modules for the tasks
